+++ Gruntfile.js
@@ -3,23 +3,28 @@ require('dotenv').config();
 module.exports = function (grunt) {
     grunt.loadNpmTasks('grunt-screeps');
     grunt.loadNpmTasks('grunt-replace');
+    grunt.loadNpmTasks('grunt-cleanempty');
     grunt.initConfig({
         screeps: {
             options: {
                 email: process.env.email,
-                password: process.env.pass,
+                token: process.env.token,
                 branch: 'default',
             },
             dist: {
                 src: ['dist/*.js'],
             },
         },
+        cleanempty: {
+            options: {},
+            src: ['dist/**/*'],
+        },
         replace: {
             dist: {
                 options: {
                     patterns: [
                         {
-                            match: /(?<=require\(".)(.*)(?=\/[\w\.]+")/g,
+                            match: /(?<=require\(["'].)(.*)(?=\/[\w\.]+["'])/g,
                             replacement: '',
                         },
                     ],
@@ -38,5 +43,5 @@ module.exports = function (grunt) {
         },
     });
 
-    grunt.registerTask('default', ['replace', 'screeps']);
+    grunt.registerTask('default', ['cleanempty', 'replace', 'screeps']);
 };
