+++ package.json
@@ -32,6 +32,7 @@
         "typescript": "^4.6.2"
     },
     "devDependencies": {
+        "grunt-cleanempty": "^1.0.4",
         "husky": "^7.0.4",
         "lint-staged": "^12.3.7",
         "prettier": "^2.6.0",
