+++ src/interfaces/room.ts
@@ -1,0 +1,13 @@
+interface RoomMemory {
+    availableSourceAccessPoints: string[];
+    sourceAccessPointCount: number;
+    roadsConstructed?: boolean;
+}
+
+interface Room {
+    initRoomMemory(): void;
+}
+
+interface RoomPosition {
+    toMemSafe(): string;
+}
