+++ src/modules/creepDriver.ts
@@ -1,5 +1,5 @@
 import { Worker } from '../roles/worker';
-import { WaveCreep } from '../types/waveCreep';
+import { WaveCreep } from './WaveCreep';
 
 export default function driveCreep(creep: Creep) {
     let waveCreep: WaveCreep;
