+++ src/prototypes/requirePrototypes.ts
@@ -1,0 +1,1 @@
+require('./room');
